+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 - Define the available types of packet in the `PacketType` enum.
 - Define the `Packet` class as the super-type of all packets.
+- Model packets:
+  - `open`: `OpenPacket`